Add Increment A-Packages
Refactor the files into 3 different packages, task contains the Task class and its subclasses, exception contains the Exception classes and duke contains the main classes
for running the program.
