+++ src/main/java/duke/UI.java
@@ -1,3 +1,7 @@
+package duke;
+
+import task.Task;
+
 import java.util.Arrays;
 
 public class UI {
