+++ src/main/java/exception/EmptyTodoException.java
@@ -1,9 +1,10 @@
-public class UnknownCommandException extends Throwable {
+package exception;
+
+public class EmptyTodoException extends Throwable{
     @Override
     public String getMessage() {
         return "____________________________________________________________\n" +
-                "I'm sorry, but I don't know what that means. :(\n" +
-                "Please enter a valid command. :)\n" +
+                "The description of a todo cannot be empty!\n" +
                 "____________________________________________________________\n";
     }
 }
