+++ src/main/java/exception/UnknownCommandException.java
@@ -1,8 +1,11 @@
-public class EmptyTodoException extends Throwable{
+package exception;
+
+public class UnknownCommandException extends Throwable {
     @Override
     public String getMessage() {
         return "____________________________________________________________\n" +
-                "The description of a todo cannot be empty!\n" +
+                "I'm sorry, but I don't know what that means. :(\n" +
+                "Please enter a valid command. :)\n" +
                 "____________________________________________________________\n";
     }
 }
